Ensure all columns are populated at ingestion.
The location table load was firing a warning because the order_id column was not
being set to a value. This is benign since the order_id values are generated in
a following sql statement, but this change ensures they are set to a default so
the warning is not shown.

Similarly the resource_organization was not being set in the fact table. This
is not actually a problem in the open source build since the organization is
not an active group by. But it should be set correctly nevertheless.