Add command line helper that runs the pipelines.
The command line helper runs the etlv2 pipelines and the filter list
builder. This also uses the journal helper to guarantee that all
data are aggregated even if the ingestor is kill prematurely.

The config settings for the pipeline are now obtained from two sources:
- The mandatory command line parameters (resource, log file path and url)
- The portal_settings (geoip path, apahce log format)

Readme and test harness updated appropriately.